requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.11
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.5
langchain-core==0.2.8
langchain-openai==0.1.14
langchain-text-splitters==0.2.1
langgraph==0.1.5
langsmith==0.1.81
loguru==0.7.3
marshmallow==3.26.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.65.4
orjson==3.10.15
packaging==24.2
propcache==0.3.0
pydantic==1.10.13
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.46.0
tenacity==8.5.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
ulid==1.1
urllib3==2.3.0
yarl==1.18.3
